requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
mango-jwt==1.3.2
passlib==1.7.4
PyJWT==2.8.0
pymongo==3.12.3
pytz==2024.1
sqlparse==0.2.4
